fix: changing numeric values on entities and mappers from integer to big decimal

+++ src/main/java/com/empresa/projeto/gestao_ubs/Dto/Category/CategoryResponseDto.java
@@ -2,6 +2,7 @@ package com.empresa.projeto.gestao_ubs.Dto.Category;
 
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -12,7 +13,7 @@ public class CategoryResponseDto {
 
     private Long category_id;
     private String name;
-    private Integer daily_limit;
-    private Integer monthly_limit;
+    private BigDecimal daily_limit;
+    private BigDecimal monthly_limit;
     private LocalDateTime created_at;
 }
