+++ src/main/java/com/empresa/projeto/gestao_ubs/Dto/Category/CategoryUpdateDto.java
@@ -2,6 +2,8 @@ package com.empresa.projeto.gestao_ubs.Dto.Category;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
@@ -9,6 +11,6 @@ import lombok.*;
 public class CategoryUpdateDto {
 
     private String name;
-    private Integer daily_limit;
-    private Integer monthly_limit;
+    private BigDecimal daily_limit;
+    private BigDecimal monthly_limit;
 }
